fix: fcm 토큰 삭제 시 발생할 수 있는 IDOR 문제 해결
- 사용자의 토큰이 맞는지 확인하는 검증로직을 추가
- 검증을 통해 해커의 토큰 탈취를 통한 별개 유저의 알림을 방해하는 공격을 막을 수 있도록 함

+++ app-main/src/main/java/net/causw/app/main/domain/user/account/service/v2/implementation/UserPushTokenWriter.java
@@ -6,6 +6,7 @@ import java.util.Set;
 import org.springframework.stereotype.Component;
 
 import net.causw.app.main.domain.user.account.entity.user.User;
+import net.causw.app.main.shared.exception.errorcode.AuthErrorCode;
 import net.causw.app.main.shared.infra.redis.RedisUtils;
 import net.causw.global.constant.StaticValue;
 
@@ -54,7 +55,10 @@ public class UserPushTokenWriter {
 	 * @param fcmToken 삭제할 FCM 토큰
 	 */
 	public void removeFcmToken(User user, String fcmToken) {
-		user.removeFcmToken(fcmToken);
+		boolean isRemoved = user.removeFcmToken(fcmToken);
+		if (!isRemoved) {
+			throw AuthErrorCode.NO_PERMISSION_FOR_RESOURCE.toBaseException();
+		}
 		redisUtils.deleteFcmTokenData(fcmToken);
 	}
